pyproject.toml
[project]
name = "taluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.8",
    "pydantic>=2.10.6",
    "surrealdb>=1.0.3",
    "uvicorn>=0.34.0",
]
